+++ src/main/java/com/casadelacultura/casadelacultura/repositorio/AudioLibroRepositorio.java
@@ -3,5 +3,5 @@ package com.casadelacultura.casadelacultura.repositorio;
 import org.springframework.data.repository.CrudRepository;
 import com.casadelacultura.casadelacultura.entity.AudioLibro;
 
-public interface AudioLibroRepositorio extends CrudRepository<AudioLibro, Integer> {
+public interface AudioLibroRepositorio extends CrudRepository<AudioLibro, Long> {
 }
