+++ src/main/java/com/casadelacultura/casadelacultura/repositorio/CategoriaLibroRepositorio.java
@@ -3,5 +3,5 @@ package com.casadelacultura.casadelacultura.repositorio;
 import org.springframework.data.repository.CrudRepository;
 import com.casadelacultura.casadelacultura.entity.CategoriaLibro;
 
-public interface CategoriaLibroRepositorio extends CrudRepository<CategoriaLibro, Integer> {
+public interface CategoriaLibroRepositorio extends CrudRepository<CategoriaLibro,Long> {
 }
