+++ src/main/java/com/casadelacultura/casadelacultura/repositorio/CategoriaObraRepositorio.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.casadelacultura.casadelacultura.entity.CategoriaObra;
 
-public interface CategoriaObraRepositorio  extends CrudRepository<CategoriaObra, Integer> {
+public interface CategoriaObraRepositorio  extends CrudRepository<CategoriaObra, Long> {
     
     
 }
