+++ src/main/java/com/casadelacultura/casadelacultura/repositorio/ConvocatoriasRepositorio.java
@@ -3,5 +3,5 @@ package com.casadelacultura.casadelacultura.repositorio;
 import org.springframework.data.repository.CrudRepository;
 import com.casadelacultura.casadelacultura.entity.Convocatorias;
 
-public interface ConvocatoriasRepositorio extends CrudRepository<Convocatorias, Integer> {
+public interface ConvocatoriasRepositorio extends CrudRepository<Convocatorias, Long> {
 }
