+++ src/main/java/com/casadelacultura/casadelacultura/repositorio/LibrosPorAutorRepositorio.java
@@ -3,5 +3,5 @@ package com.casadelacultura.casadelacultura.repositorio;
 import org.springframework.data.repository.CrudRepository;
 import com.casadelacultura.casadelacultura.entity.LibrosPorAutor;
 
-public interface LibrosPorAutorRepositorio extends CrudRepository<LibrosPorAutor, Integer> {
+public interface LibrosPorAutorRepositorio extends CrudRepository<LibrosPorAutor, Long> {
 }
