+++ src/main/java/com/casadelacultura/casadelacultura/repositorio/MaterialRepositorio.java
@@ -3,5 +3,5 @@ package com.casadelacultura.casadelacultura.repositorio;
 import org.springframework.data.repository.CrudRepository;
 import com.casadelacultura.casadelacultura.entity.Material;
 
-public interface MaterialRepositorio extends CrudRepository<Material, Integer> {
+public interface MaterialRepositorio extends CrudRepository<Material, Long> {
 }
